+++ Library/Aphid/Support/Compatibility/Date.js
@@ -1,0 +1,11 @@
+/**
+ * Aphid.Support.Compatibility.Date
+**/
+
+if (Object.isUndefined(Date.now))
+{
+  Date.now = function()
+  {
+    return new Date().getTime();
+  }
+}